+++ Timelapse2/Timelapse/Images/Marker.cs
@@ -61,7 +61,7 @@ namespace Timelapse.Images
         // Calculate the point as a ratio of its position on the image, so we can locate it regardless of the actual image size
         public static Point ConvertPointToRatio(Point p, double width, double height)
         {
-            Point ratio = new Point((double)p.X / (double)width, (double)p.Y / (double)height);
+            Point ratio = new Point(p.X / width, p.Y / height);
             return ratio;
         }
